python-bot/requirements.txt
telegram
python-telegram-bot
gspread 
oauth2client
Pillow
pytesseract
psycopg2
asyncio